backend/requirements.txt
# Core
Django>=5.1,<5.2
djangorestframework>=3.15,<4.0
django-cors-headers>=4.4,<5.0

# Database
psycopg2-binary>=2.9,<3.0

# Authentication
dj-rest-auth[with_social]>=7.0,<8.0
djangorestframework-simplejwt>=5.3,<6.0
django-allauth>=65.0,<66.0

# Cache
django-redis>=5.4,<6.0

# WebSocket / Realtime
channels[daphne]>=4.0,<5.0
channels-redis>=4.2,<5.0

# Utilities
python-dotenv>=1.0,<2.0
django-filter>=24.0,<25.0
icalendar>=6.0,<7.0
requests>=2.32,<3.0

# Celery / async tasks
celery[redis]>=5.3,<6.0
django-celery-beat>=2.6,<3.0

# Recurring events
python-dateutil>=2.9,<3.0

# Development
pytest-django>=4.8,<5.0
factory-boy>=3.3,<4.0
